part5 chapter 6

+++ src/main/java/org/zerock/controller/UploadController.java
@@ -1,7 +1,15 @@
 package org.zerock.controller;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.InputStream;
+import java.util.UUID;
+
+import javax.annotation.Resource;
+
 import org.apache.commons.io.IOUtils;
-import org.slf4j.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -11,90 +19,86 @@ import org.springframework.ui.Model;
 import org.springframework.util.FileCopyUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.zerock.util.MediaUtils;
 import org.zerock.util.UploadFileUtils;
 
-import javax.annotation.Resource;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStream;
-import java.util.UUID;
-
 
 @Controller
 public class UploadController {
 
-    private static final Logger logger =
-            LoggerFactory.getLogger(UploadController.class);
+    private static final Logger logger = LoggerFactory.getLogger(UploadController.class);
 
-    @Resource(name ="uploadPath")
+    @Resource(name = "uploadPath")
     private String uploadPath;
 
     @RequestMapping(value = "/uploadForm", method = RequestMethod.GET)
-    public void uploadForm(){
-
-
-
+    public void uploadForm() {
     }
 
     @RequestMapping(value = "/uploadForm", method = RequestMethod.POST)
-    public String  uploadForm(MultipartFile file, Model model) throws Exception{
-        logger.info("originalName : " + file.getOriginalFilename());
-        logger.info("size : " + file.getSize());
-        logger.info("contentType : " + file.getContentType());
+    public String uploadForm(MultipartFile file, Model model) throws Exception {
 
-        String savedName =
-                uploadFile(file.getOriginalFilename(), file.getBytes());
+        logger.info("originalName: " + file.getOriginalFilename());
+        logger.info("size: " + file.getSize());
+        logger.info("contentType: " + file.getContentType());
+
+        String savedName = uploadFile(file.getOriginalFilename(), file.getBytes());
+
         model.addAttribute("savedName", savedName);
 
         return "uploadResult";
     }
 
-    private String uploadFile(String originalFilename, byte[] fileData)throws Exception {
-        UUID uid = UUID.randomUUID();
-
-        String savedName = uid.toString() + "_" + originalFilename;
-
-        File target = new File(uploadPath,savedName);
-
-        FileCopyUtils.copy(fileData,target);
-
-        return savedName;
+    @RequestMapping(value = "/uploadAjax", method = RequestMethod.GET)
+    public void uploadAjax() {
     }
 
+    private String uploadFile(String originalName, byte[] fileData) throws Exception {
 
-    @RequestMapping(value = "uploadAjax", method = RequestMethod.GET)
-    public void uplaodAjax(){}
+        UUID uid = UUID.randomUUID();
+
+        String savedName = uid.toString() + "_" + originalName;
+
+        File target = new File(uploadPath, savedName);
+
+        FileCopyUtils.copy(fileData, target);
+
+        return savedName;
+
+    }
 
     @ResponseBody
-    @RequestMapping(value = "uploadAjax", method = RequestMethod.POST
-                    ,produces = "text/plain;charset=UTF-8")
-    public ResponseEntity<String> uploadAjax(MultipartFile file) throws  Exception{
-        logger.info("originalName : " + file.getOriginalFilename());
-        logger.info("size : " + file.getSize());
-        logger.info("contentType : " + file.getContentType());
+    @RequestMapping(value ="/uploadAjax", method=RequestMethod.POST,
+            produces = "text/plain;charset=UTF-8")
+    public ResponseEntity<String> uploadAjax(MultipartFile file)throws Exception{
+
+        logger.info("originalName: " + file.getOriginalFilename());
+
 
         return
                 new ResponseEntity<>(
-                        UploadFileUtils.uploadFile(
-                                uploadPath,
+                        UploadFileUtils.uploadFile(uploadPath,
                                 file.getOriginalFilename(),
                                 file.getBytes()),
                         HttpStatus.CREATED);
     }
 
+
     @ResponseBody
-    @RequestMapping(value = "/displayFile")
-    public ResponseEntity<byte[]> displayFile(String fileName)throws Exception{
+    @RequestMapping("/displayFile")
+    public ResponseEntity<byte[]>  displayFile(String fileName)throws Exception{
+
         InputStream in = null;
         ResponseEntity<byte[]> entity = null;
 
-        logger.info("FILE NAME : " + fileName);
+        logger.info("FILE NAME: " + fileName);
 
         try{
-            String formatName = fileName.substring(fileName.lastIndexOf('.')+1);
+
+            String formatName = fileName.substring(fileName.lastIndexOf(".")+1);
 
             MediaType mType = MediaUtils.getMediaType(formatName);
 
@@ -102,47 +106,74 @@ public class UploadController {
 
             in = new FileInputStream(uploadPath+fileName);
 
-            if(mType != null)
+            if(mType != null){
                 headers.setContentType(mType);
-            else {
+            }else{
+
                 fileName = fileName.substring(fileName.indexOf("_")+1);
                 headers.setContentType(MediaType.APPLICATION_OCTET_STREAM);
                 headers.add("Content-Disposition", "attachment; filename=\""+
-                    new String(fileName.getBytes("UTF-8"), "ISO-8859-1")+"\"");
-
+                        new String(fileName.getBytes("UTF-8"), "ISO-8859-1")+"\"");
             }
-            entity = new ResponseEntity<byte[]>(IOUtils.toByteArray(in),headers,HttpStatus.CREATED);
-        }catch (Exception e){
+
+            entity = new ResponseEntity<byte[]>(IOUtils.toByteArray(in),
+                    headers,
+                    HttpStatus.CREATED);
+        }catch(Exception e){
             e.printStackTrace();
             entity = new ResponseEntity<byte[]>(HttpStatus.BAD_REQUEST);
-        }finally {
+        }finally{
             in.close();
         }
         return entity;
     }
 
     @ResponseBody
-    @RequestMapping(value = "/deleteFile", method = RequestMethod.POST)
+    @RequestMapping(value="/deleteFile", method=RequestMethod.POST)
     public ResponseEntity<String> deleteFile(String fileName){
-        logger.info("delete file : " + fileName );
+
+        logger.info("delete file: "+ fileName);
 
         String formatName = fileName.substring(fileName.lastIndexOf(".")+1);
 
         MediaType mType = MediaUtils.getMediaType(formatName);
 
         if(mType != null){
+
             String front = fileName.substring(0,12);
             String end = fileName.substring(14);
-            new File(uploadPath + (front+end).replace('/',File.separatorChar)).delete();
-
-
+            new File(uploadPath + (front+end).replace('/', File.separatorChar)).delete();
         }
-        new File(uploadPath + fileName.replace('/',File.separatorChar)).delete();
-        return new ResponseEntity<>("deleted",HttpStatus.OK);
+
+        new File(uploadPath + fileName.replace('/', File.separatorChar)).delete();
+
+
+        return new ResponseEntity<String>("deleted", HttpStatus.OK);
     }
 
+    @ResponseBody
+    @RequestMapping(value = "/deleteAllFiles", method=RequestMethod.POST)
+    public ResponseEntity<String> deleteFile1(@RequestParam("files[]") String[] files){
+        logger.info("delete all files: "+files);
 
+        if(files == null || files.length==0) {
+            return new ResponseEntity<String>("deleted", HttpStatus.OK);
+        }
 
+        for(String fileName : files) {
+            String formatName = fileName.substring(fileName.lastIndexOf(".")+1);
 
+            MediaType mType = MediaUtils.getMediaType(formatName);
+
+            if(mType !=null) {
+
+                String front = fileName.substring(0,12);
+                String end = fileName.substring(14);
+                new File(uploadPath + (front+end).replace('/', File.separatorChar)).delete();
+            }
+            new File(uploadPath+fileName.replace('/',File.separatorChar)).delete();
+        }
+        return new ResponseEntity<String>("deleted",HttpStatus.OK);
+    }
 
 }
